Monad Mainnet Boosted Markets

+++ contracts/contracts/pancakev3/interfaces/IPancakeV3Pool.sol
@@ -11,11 +11,19 @@ interface IPancakeV3Pool {
             uint16 observationIndex,
             uint16 observationCardinality,
             uint16 observationCardinalityNext,
-            uint8 feeProtocol,
+            uint32 feeProtocol, // Note: Monad uses uint32, not uint8
             bool unlocked
         );
 
     function tickSpacing() external view returns (int24);
 
-    function observe(uint32[] calldata secondsAgos) external view returns (int56[] memory tickCumulatives, uint160[] memory secondsPerLiquidityCumulativeX128s);
+    function observe(
+        uint32[] calldata secondsAgos
+    )
+        external
+        view
+        returns (
+            int56[] memory tickCumulatives,
+            uint160[] memory secondsPerLiquidityCumulativeX128s
+        );
 }
